Converted the JDK collection test library + list tests
Converted a number of List tests to TypeScript (those that don't require classes not in the JREE), which required to add full support for annotations and an emulation of some TestNG test library emulation. This is still ongoing work, but does at least compile now. Details:

- Added AtomicBoolean, AtomicInteger, ClassCastException, StackoverflowError, IntBinaryOperator and IntUnaryOperator classes/interfaces.
- Enhanced Boolean, Integer and Long comparison methods to handle also primitive values.
- Changed all functional interfaces to use a call signature, which makes it easier using closures as such an interface. No need for the creation functions anymore.
- Started adding null pointer exceptions for null parameters in certain places. More to come probably.
- Added a number of methods in some classes which are used in the tests.
- Made java.util.Comparator a real functional interface (alike).
- Added decorators for org.testng.annotation.Test and .DataProvider.

Note: the new tests are not yet executed.