The ListDefaults JDK tests succeed now.
This test suite tests many standard things in various lists (including Vector, ArrayList etc.).

This patch lays the ground for the execution of JDK tests in general, by implementing some of the TestNG functionality. That included extending the test decorators to support more parameters (description. enabled, timeout) and support of data providers.

For better integration with Jest an own matcher was added to compare Java objects.